+++ src/apis/api/index.ts
@@ -1,0 +1,35 @@
+import { postAuthJoin, getAuthLogin, getAuthIsCheck } from './auth'
+import {
+  getUserInfo,
+  deleteUser,
+  postUserInfo,
+  postUserSites,
+  deleteUserSites,
+  postUserCheck,
+  postUserPassword,
+} from './user'
+import { getSites } from './site'
+import { getContents, getContentsOfSite } from './content'
+
+export {
+  //auth
+  postAuthJoin,
+  getAuthLogin,
+  getAuthIsCheck,
+
+  //user
+  getUserInfo,
+  deleteUser,
+  postUserInfo,
+  postUserSites,
+  deleteUserSites,
+  postUserCheck,
+  postUserPassword,
+
+  //site
+  getSites,
+
+  //content
+  getContents,
+  getContentsOfSite,
+}
