+++ src/layouts/EditUserInfoList.tsx/Service.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { useMutation, useQuery } from 'react-query'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import { userInfoState } from '../../recoil/userInfoState'
-import { deleteUserSites, getSites, postUserSites } from '../../service/apis'
+import { deleteUserSites, getSites, postUserSites } from '../../apis/api'
 import { TSites } from '../../types/commonTypes'
 import EditService from '../EditService'
 import SavedService from '../SavedService'
