+++ src/view/MyPage.tsx
@@ -5,7 +5,7 @@ import { useLogout } from '../hooks/useLogout'
 import { userInfoState } from '../recoil/userInfoState'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import { useMutation } from 'react-query'
-import { deleteUser } from '../service/apis'
+import { deleteUser } from '../apis/api'
 import { useNavigate } from 'react-router-dom'
 import SavedService from '../layouts/SavedService'
 import { modalState } from '../recoil/useModalState'
