+++ src/view/Register.tsx
@@ -8,7 +8,7 @@ import {
 } from '@heroicons/react/24/solid'
 import Header from '../layouts/Header'
 import { useMutation, useQuery } from 'react-query'
-import { getSites, getAuthIsCheck, postAuthJoin } from '../service/apis'
+import { getSites, getAuthIsCheck, postAuthJoin } from '../apis/api'
 import Loading from '../components/Loading'
 import { useSetUserInfo } from '../hooks/useSetUserInfo'
 import { TUserInfo, TSites } from '../types/commonTypes'
